+++ src/config/mode-registry.ts
@@ -1,4 +1,4 @@
-import { Config } from '../config/config';
+import { Config } from './config';
 
 
 export const MODE_IOS: any = {
